+++ src/main/java/com/season/livingmate/auth/dto/response/RefreshTokenResDto.java
@@ -1,9 +1,6 @@
 package com.season.livingmate.auth.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.season.livingmate.user.entity.User;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
