+++ src/main/java/com/season/livingmate/auth/dto/response/SignupResDto.java
@@ -1,6 +1,6 @@
 package com.season.livingmate.auth.dto.response;
 
-import com.season.livingmate.user.entity.User;
+import com.season.livingmate.user.domain.User;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
