+++ src/main/java/com/season/livingmate/auth/service/LogoutService.java
@@ -7,12 +7,11 @@ import com.season.livingmate.auth.security.JwtProvider;
 import com.season.livingmate.exception.Response;
 import com.season.livingmate.exception.status.ErrorStatus;
 import com.season.livingmate.exception.status.SuccessStatus;
-import com.season.livingmate.user.entity.User;
-import com.season.livingmate.user.repository.UserRepository;
+import com.season.livingmate.user.domain.User;
+import com.season.livingmate.user.domain.repository.UserRepository;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
