+++ src/main/java/com/season/livingmate/auth/service/SignupService.java
@@ -6,8 +6,8 @@ import com.season.livingmate.auth.dto.response.SignupResDto;
 import com.season.livingmate.auth.security.JwtProvider;
 import com.season.livingmate.exception.CustomException;
 import com.season.livingmate.exception.status.ErrorStatus;
-import com.season.livingmate.user.entity.User;
-import com.season.livingmate.user.repository.UserRepository;
+import com.season.livingmate.user.domain.User;
+import com.season.livingmate.user.domain.repository.UserRepository;
 import jakarta.annotation.PostConstruct;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
