+++ src/main/java/com/season/livingmate/user/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.season.livingmate.user.repository;
+package com.season.livingmate.user.domain.repository;
 
-import com.season.livingmate.user.entity.User;
+import com.season.livingmate.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
